requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# electracker/src/export_img.py: 13
Pillow == 8.1.0

# electracker/src/violin/présidentielles.py: 10
kaleido == 0.1.0

# electracker/src/tweet_chiffres_quotidiens.py: 6
# electracker/src/violin/présidentielles.py: 5,9
# electracker/traiter_resultats_elections.py: 6
# electracker/traiter_sondages_elections.py: 6
numpy == 1.21.4

# electracker/src/elections_presidentielles_moyenne_mobile_intentions.py: 2
# electracker/src/violin/présidentielles.py: 3
# electracker/traiter_resultats_elections.py: 5
# electracker/traiter_sondages_elections.py: 5
pandas == 1.2.1

# electracker/src/tweet_chiffres_quotidiens.py: 3
# electracker/src/violin/présidentielles.py: 4,8
plotly == 5.4.0

# electracker/src/violin/présidentielles.py: 6
scipy == 1.6.0

# electracker/src/export_img.py: 2
selenium == 4.1.0

# electracker/src/tweet_chiffres_quotidiens.py: 1
# electracker/src/violin/présidentielles.py: 12
tweepy == 3.10.0

# electracker/src/export_img.py: 5
webdriver_manager == 3.5.2
